Make ipfs webui available at `freenas_ip/ipfsui`.
I spent a day figuring out why nginx did not like me!

Turns out I am stupid, s'all.

Still need to fix its daemonConf so that the above page loads with proper information.

Right now it loads the correct page but the releavnt ipfs info is not present on it.