Make sure any subprocess.Popen python call's do not close 4 million (or ulimit -n) file descriptors when close_fds=True.
This only helps in python3.4 and not python2.7 but still...

We are going to make a habbit of changing any python popen calls that we make have the
close_fds=False set but still 3rd party libraries and such making said calls cannot be controlled.

This fix ensures that only the fd's in use will be closed and not close from 1->max_fds!

Here is the brief paragraph from the freebsd mailing list that I found on this:
"""
Well the python devs are aware by themselves of potential performances issues on
FreeBSD (and non linux in general) for example subprocess will try to close fds,
on linux by getting the list of fd from /proc/fd and only close the one they do
not want among the existing ones. on freebsd they do the same if /dev/fd is
mounted meaning without /dev/fd, perfs will suck. They do not use closefrom(2)
here because on linux it is not async-signal-safe. one could make them use
closefrom(2) on non linux for example or even more efficiently but freebsd only
modify the code to use kinfo_getfile(3).
"""
For more info: http://lists.freebsd.org/pipermail/freebsd-virtualization/2015-November/003936.html